moved BitField into Substrate, no tests yet
