+++ StarFire/src/StarFire.h
@@ -1,9 +1,11 @@
 #pragma once
 
-#include "StarFire/Core/Logging.h"
-#include "StarFire/Core/Core.h"
+//Core
 #include "StarFire/Core/Application.h"
-#include "StarFire/Core/Timestep.h"
-
+#include "StarFire/Core/Core.h"
+#include "StarFire/Events/Event.h"
 #include "StarFire/Core/Layer.h"
+#include "StarFire/Core/Logging.h"
 
+
+#include "StarFire/Core/Timestep.h"
